Enhanced project structure
