+++ src/components/areas/Footer.jsx
@@ -1,0 +1,5 @@
+export default function Footer() {
+    return (
+        <h2>Footer</h2>
+    )
+}
