+++ src/components/areas/Main.jsx
@@ -1,0 +1,13 @@
+import Projects from "../sections/Projects.jsx";
+import Experience from "../sections/Experience.jsx";
+import AboutMe from "../sections/AboutMe.jsx";
+
+export default function Main() {
+    return (
+        <>
+            <AboutMe />
+            <Projects />
+            <Experience />
+        </>
+    )
+}
