+++ src/components/sections/AboutMe.jsx
@@ -1,0 +1,7 @@
+export default function AboutMe() {
+    return (
+        <section>
+            <h2>AboutMe</h2>
+        </section>
+    )
+}
