+++ src/components/sections/Projects.jsx
@@ -1,0 +1,7 @@
+export default function Projects() {
+    return (
+        <section>
+            <h2>Projects</h2>
+        </section>
+    )
+}
